Proyecto_Netflix_Python/requirements.txt
pyodbc
SQLAlchemy
python-dotenv
bcrypt